LayoutTests/fast/regions/webkit-named-flow-event-target-expected.txt

Tests the the NamedFlow object behavior as an EventTarget

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Flow content
PASS event.target.name is "article"
PASS getFlowByName("article") is non-null.
PASS recreated is false
PASS getFlowByName("article") is null
PASS event.target.name is "article"
PASS recreated is true
PASS getFlowByName("article") is non-null.
PASS updatedCount is 2
PASS successfullyParsed is true

TEST COMPLETE
 
